+++ packages/guide/src/app/shared/services/database.service.ts
@@ -107,6 +107,8 @@ export class DatabaseService {
                         this.fetchShopProcessItems$("lab"),
                         this.fetchShopItemData$("carpenter"),
                         this.fetchShopItemData$("general-store"),
+                        this.fetchShopItemData$("ranch"),
+                        this.fetchShopItemData$("beach-shack"),
                         this.fetchShopItemData$("merfolk-general-store"),
                         this.fetchShopItemData$("merfolk-oracle-tail-store"),
                         this.fetchShopItemData$("pet-shop"),